+++ tests/lib/prepare.js
@@ -1,8 +1,8 @@
-commonDiv = null;
-Meteor.startup(function() {
-  commonDiv = document.createElement('div');
-  commonDiv.style.display = 'none';
-  document.body.appendChild(commonDiv);
-});
-
 noop = function() {};
+
+testDiv = function() {
+  var div = document.createElement('div');
+  div.style.display = 'none';
+  document.body.appendChild(div);  
+  return div;
+}